requirements.txt
fastapi>=0.115,<1.0
docxtpl>=0.16,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.6,<3.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29,<1.0
greenlet>=3.0
alembic>=1.13,<2.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0
ruff>=0.8,<1.0
redis[asyncio]>=5.0,<6.0
